refactor: 스케줄링 시간 수정

+++ src/main/java/in/koreatech/koin/domain/club/repository/redis/ClubHitsRedisRepository.java
@@ -23,7 +23,7 @@ public class ClubHitsRedisRepository {
 
     public Long getHits(String key) {
         Object value = redisTemplate.opsForValue().get(key);
-        return value instanceof Long ? (Long) value : 0L;
+        return value instanceof Integer ? (Integer) value : 0L;
     }
 
     public void deleteHitsByKeys(Set<String> keys) {
